+++ src/refactor_nrepl/ns/prune_dependencies.clj
@@ -1,9 +1,9 @@
 (ns refactor-nrepl.ns.prune-dependencies
   (:require
    [cider.nrepl.middleware.info :as info]
-   [refactor-nrepl.config :as config]
    [refactor-nrepl.core :as core]
    [refactor-nrepl.find.symbols-in-file :as symbols-in-file]
+   [refactor-nrepl.ns.libspec-allowlist :as libspec-allowlist]
    [refactor-nrepl.util :as util]))
 
 (defn- lookup-symbol-ns
@@ -108,11 +108,13 @@
 ;; Some namespaces, e.g. those containing only protocol extensions,
 ;; are side-effecting at load but might look unused and otherwise be
 ;; pruned.
-(defn- libspec-should-never-be-pruned? [libspec]
+(defn libspec-should-never-be-pruned?
+  "Should `libspec` never be pruned away by the `clean-ns` op?"
+  [libspec]
   (let [ns-name (str (:ns libspec))]
-    (some (fn [^String pattern]
-            (re-find (re-pattern pattern) ns-name))
-          (:libspec-whitelist config/*config*))))
+    (boolean (some (fn [^String pattern]
+                     (-> pattern re-pattern (re-find ns-name)))
+                   (libspec-allowlist/libspec-allowlist)))))
 
 (defn- prune-libspec [symbols-in-file current-ns libspec]
   (if (libspec-should-never-be-pruned? libspec)